requirements.txt
ciscoconfparse
openpyxl
scrapli
rich
python-dotenv
pandas
tqdm
scrapli
scrapli-community